producer_and_consumer_scripts/requirements.txt
math
pandas
shapely
confluent-kafka
pandas-gbq
google-cloud-bigquery
google-auth
PyYAML==5.4.1